+++ Warehouses/src/main/resources/db/migration/V4__create_shop_warehouse_relationship.sql
@@ -1,6 +1,6 @@
 CREATE TABLE shop_warehouse (
-    shop_id BIGINT NOT NULL,
-    warehouse_id BIGINT NOT NULL,
+    shop_id INTEGER NOT NULL,
+    warehouse_id INTEGER NOT NULL,
 
     PRIMARY KEY (shop_id, warehouse_id),
